pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llamaindex-nativ"
version = "0.1.0"
description = "LlamaIndex integration for Nativ — AI-powered localization tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Nativ", email = "hello@usenativ.com" }]
keywords = ["llamaindex", "llama-index", "localization", "translation", "i18n", "l10n", "ai", "nativ"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Localization",
    "Typing :: Typed",
]
dependencies = [
    "nativ>=0.1.0",
    "llama-index-core>=0.11.0",
]

[project.urls]
Homepage = "https://usenativ.com"
Documentation = "https://docs.usenativ.com"
Repository = "https://github.com/Nativ-Technologies/llamaindex-nativ"
Issues = "https://github.com/Nativ-Technologies/llamaindex-nativ/issues"

[tool.hatch.build.targets.wheel]
packages = ["llamaindex_nativ"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]
